% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentTaxonomy.R
\name{augmentTaxonomy}
\alias{augmentTaxonomy}
\title{Augment a taxonomy table}
\usage{
augmentTaxonomy(
  taxtab,
  useBinomSpecies = TRUE,
  numberUnclassified = TRUE,
  replaceIncertaeSedis = TRUE
)
}
\arguments{
\item{taxtab}{taxonomy table, e.g. as produced by the dada2 pipeline}

\item{useBinomSpecies}{replaces a species name by Genus + species name, if present (default = TRUE)}

\item{numberUnclassified}{use numbers to distinguish uncharacterized taxa that belong to the same higher-level taxon (default = TRUE)}

\item{replaceIncertaeSedis}{if any name is "Incertae Sedis", this name is replaced by the next higher tax level with the addition "_i.s." (default = TRUE)}
}
\value{
augmented taxonomy table
}
\description{
Uncharacterized taxa (name "NA") will be replaced with the names of the last defined taxonomic level.
The taxa can additionally be numbered
}
\examples{
taxtab <- augmentTaxonomy(taxtab)
taxtab <- augmentTaxonomy(taxtab, numberUnclassified = F)
}
