% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_bargraph_top_taxa.R, R/bargraph_top_taxa.R
\name{bargraph_top_taxa}
\alias{bargraph_top_taxa}
\title{Phyloseq-style bargraph of the top k taxa}
\usage{
bargraph_top_taxa(
  physeq,
  k = 17,
  taxlevel = "Genus",
  relative = T,
  groupBy = NULL,
  groupBy2 = NULL,
  mergelevel = F,
  colorfill = F,
  colorPalette = NULL,
  othercol = "#BBBBBB",
  textsize = 10,
  maxlegend = 9,
  xlabel = NULL
)

bargraph_top_taxa(
  physeq,
  k = 17,
  taxlevel = "Genus",
  relative = T,
  groupBy = NULL,
  groupBy2 = NULL,
  mergelevel = F,
  colorfill = F,
  colorPalette = NULL,
  othercol = "#BBBBBB",
  textsize = 10,
  maxlegend = 9,
  xlabel = NULL
)
}
\arguments{
\item{physeq}{A phyloseq object containing your microbiome data set.}

\item{k}{Number of most abundant taxa that should be displayed. Defaults to 17 (corresponding to the default palette.)}

\item{taxlevel}{Taxonomic level used for coloring the graph. Must use only levels that are defined in tax_table(physeq). Defaults to Genus.}

\item{relative}{If TRUE, use relative abundances, else use absolute counts. Default: TRUE.}

\item{groupBy}{(optional) If used, uses this variable to group samples and display sums/average values. Must use variables that are present in sample_data(physeq).}

\item{groupBy2}{(optional) Defines a secondary grouping variable.}

\item{mergelevel}{(optional) taxa are merged on the specified level, displaying only one filled bar per taxon. Always sets colorfill = TRUE.}

\item{colorfill}{(deprecated, use mergelevel instead) If TRUE, margins of taxa bars are colored as well, displaying the whole taxon on the definedtaxlevel in one color. If FALSE, black margins will allow to distinguish low level taxa (ASVs). Default: FALSE.}

\item{colorPalette}{Define a color palette to use for the graph. Default uses RColorBrewer and combines two palettes: c(brewer.pal(9,"Set1"),brewer.pal(8,"Dark2")).}

\item{othercol}{Defines color of "other" bar that sums up all taxa beyond the k most abundant. Set to NULL to disable this bar. Default color is #BBBBBB (medium gray).}

\item{textsize}{Sets the main text size of the graph. Default is 10.}

\item{maxlegend}{Maximum number of taxa in legend before legend has two columns (Default: 9)}

\item{xlabel}{Label to be displayed below X axis (Default: NULL)}
}
\description{
Uses phyloseq bargraphs to display a defined number of most abundant taxa in the data set.

Uses phyloseq bargraphs to display a defined number of most abundant taxa in the data set.
}
\examples{
bargraph_top_taxa(physeq) #simple bargraph of all samples

bargraph_top_taxa(physeq, k=10, groupBy = "group") # bargraph of top 10 taxa, grouped by variable "group"
bargraph_top_taxa(physeq, k=10, colorfill = T, groupBy = "group") # same with colors filled

bargraph_top_taxa(physeq, k=6, colorfill = T, taxlevel = "Phylum", othercol = NULL) # top 6 phylo (no "other")
bargraph_top_taxa(physeq) #simple bargraph of all samples

bargraph_top_taxa(physeq, k=10, groupBy = "group") # bargraph of top 10 taxa, grouped by variable "group"
bargraph_top_taxa(physeq, k=10, colorfill = T, groupBy = "group") # same with colors filled

bargraph_top_taxa(physeq, k=6, colorfill = T, taxlevel = "Phylum", othercol = NULL) # top 6 phylo (no "other")
}
\seealso{
phyloseq::plot_bar()

phyloseq::plot_bar()
}
