% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_dada2_concat_pipeline.R,
%   R/dada2_concat_pipeline.R
\name{dada2_concat_pipeline}
\alias{dada2_concat_pipeline}
\title{Unified pipeline for analyzing data with different V-regions in dada2}
\usage{
dada2_concat_pipeline(
  fastq_path,
  samplefile_path,
  out_prefix,
  filter_path = "fastq_filtered/",
  silva_path = "/data/db/SILVA/dada2/silva_nr_v138_train_set.fa.gz",
  trimf = 150,
  trimr = 150,
  trimleft = 0,
  justConcatenate = T,
  maxEE = 1,
  onlyNfirst = NA
)

dada2_concat_pipeline(
  fastq_path,
  samplefile_path,
  out_prefix,
  filter_path = "fastq_filtered/",
  silva_path = "/data/db/SILVA/dada2/silva_nr_v138_train_set.fa.gz",
  trimf = 150,
  trimr = 150,
  trimleft = 0,
  justConcatenate = T,
  maxEE = 1,
  onlyNfirst = NA
)
}
\arguments{
\item{fastq_path}{Input folder with fastq read data}

\item{samplefile_path}{Input file with sample names and corresponding R1 and R2 read files}

\item{out_prefix}{Prefix for output files}

\item{filter_path}{Output path for filtered and trimmed fastq files (default: fastq_filtered/)}

\item{silva_path}{Path to SILVA database}

\item{trimf}{Fixed trimming length for R1 (default: 150)}

\item{trimr}{Fixed trimming length for R2 (default: 150)}

\item{trimleft}{trim at the 5' end of reads (e.g., to remove primers, default: 0)}

\item{justConcatenate}{Don't merge denoised R1 and R2 reads but concatenate them. (Default: TRUE)}

\item{maxEE}{max. expected error rate (per 100 basepairs of trimmed sequence) used for filtering (default: 1)}

\item{onlyNfirst}{only for testing purpose. If set to a number n, only the first n samples will be included in the analysis.}
}
\description{
Perform trimming&filtering, denoising, merging, chimera removal and taxonomic assignment steps of dada2.
If regions include amplicons larger than trimf+trimr, the justConcatenate option should be set.
Otherwise, most reads will not be merged resulting in poor results.

Perform trimming&filtering, denoising, merging, chimera removal and taxonomic assignment steps of dada2.
If regions include amplicons larger than trimf+trimr, the justConcatenate option should be set.
Otherwise, most reads will not be merged resulting in poor results.
}
