% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyPCoA.R
\name{easyPCoA}
\alias{easyPCoA}
\title{A wrapper for simplified use of phyloseq's plot_ordination}
\usage{
easyPCoA(
  physeq,
  ord,
  color = NULL,
  shape = NULL,
  size = 3,
  axes = c(1, 2),
  colorpal = NULL,
  title = NULL,
  textsize = 20,
  legend.key.size = 8
)
}
\arguments{
\item{physeq}{a phyloseq object containing your microbiome data set.}

\item{ord}{a ordination object, output of phyloseq::ordinate()}

\item{color}{name of the variable to be displayed as colors (optional)}

\item{shape}{name of the variable to be displayed as colors (optional)}

\item{size}{data point size (default: 3)}

\item{axes}{two-element vector selecting the axes to be displayed (default: c(1,2))}

\item{colorpal}{color palette used for color coding}

\item{title}{title for the plot (optional)}

\item{textsize}{base size for text elements (default: 20)}

\item{legend.key.size}{size of legend key symbols, unit: mm (default: 8)}
}
\description{
Instead of adding the same old ggplot options over and over, this function sets some options per default.
Also the last layer of the plot is removed, preventing double plotting of data points, which is common to plot_ordination.
The following ggplot options are used by default:
\itemize{
\item coord_fixed() # setting same scale to both axes
\item geom_point(size = size) # allowing control of data point sizes per parameter (see below)
\item theme(text = element_text(size = textsize), legend.key.size = unit(legend.key.size, "mm")) # allowing control of textsize and legend key size per parameter
\item scale_color_manual(values = colorpal) # setting color paletter per parameter
\item ggtitle(title) # setting the title per parameter
}
}
\examples{
myPCoA(physeq_BAL, ord_pcoa_bc, color = "feed", shape = "sampleType", colorpal = dietPalette, title = "PCoA: Bray Curtis")

}
