% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopTaxa.R
\name{getTopTaxa}
\alias{getTopTaxa}
\title{Retrieve k top taxa from a phyloseq object}
\usage{
getTopTaxa(physeq, k, taxlevel = "Genus", relative = TRUE, agglomerate = TRUE)
}
\arguments{
\item{physeq}{Phyloseq object containing the microbiome data}

\item{k}{number of most abundant taxa to retrieve}

\item{taxlevel}{taxonomic level at which the taxa should be counted and sorted (default: Genus)}

\item{relative}{use relative abundances for counting taxa (default: TRUE)}

\item{agglomerate}{sum up all lower level members of each taxon (otherwise only the most abundant sequence variant will be counted (default: TRUE)))}
}
\value{
List of names of the most abundant taxa
}
\description{
Sorts all taxa (at a specified level) by overall abundance and returns the names of the k most abundant taxa.
}
\examples{
top20genera <- getTopTaxa(physeq_feces, 20)
top5phyla  <- getTopTaxa(physeq_feces, 5, taxlevel = "Phylum")
}
