% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locatePrimers.R
\name{locatePrimers}
\alias{locatePrimers}
\title{Locates primers sequences within fastq reads}
\usage{
locatePrimers(fwdPrimer, revPrimer, fwdFiles, revFiles, onlyNfirst = NA)
}
\arguments{
\item{fwdPrimer}{forward primer sequence}

\item{revPrimer}{reverse primer sequence}

\item{fwdFiles}{fwd FASTQ filenames}

\item{revFiles}{rev FASTQ filenames}

\item{onlyNfirst}{only analyse the N first files (for testing purpose)}
}
\value{
locatedPrimers
}
\description{
This function to identify primers was adapted from the dada2 ITS pipeline workflow (https://benjjneb.github.io/dada2/ITS_workflow.html).
Thanks to benjjneb!
}
