% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapColor.R
\name{mapColor}
\alias{mapColor}
\title{Maps values to colors}
\usage{
mapColor(values, limColors, steps, lims = NULL)
}
\arguments{
\item{values}{Vector of values}

\item{lims}{two-element vector of lower and upper limit. Defaul: c(min(values), max(values))}

\item{map}{Colormap as defined e.g. by RColorBrewer}
}
\value{
color map
}
\description{
Values are converted into color statements that can e.g. be exported for use in external programs.
}
