% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSingleTaxon.R
\name{plotSingleTaxon}
\alias{plotSingleTaxon}
\title{Display only a single taxon from a phyloseq object}
\usage{
plotSingleTaxon(
  physeq,
  taxname,
  param1,
  param2 = NULL,
  taxlevel = "ASV",
  mode = "violin",
  colorpal,
  relative = TRUE
)
}
\arguments{
\item{physeq}{a phyloseq object containing your microbiome data set.}

\item{taxname}{name of the taxon to be displayed (should appear in tax_table(physeq))}

\item{param1}{grouping variable (will be used for the x axis)}

\item{param2}{(optional) secondary grouping variable (will be displayed as different colors)}

\item{taxlevel}{name of the taxonomic level at which the taxon is characterized (Default: ASV)}

\item{mode}{Which type of graph should be used? "violin" - violin plot (default), "boxplot" - boxplot, "point" - dot plot}

\item{colorpal}{Define a color palette to use for the graph.}

\item{relative}{Display relative abundances if TRUE, absolute counts if FALSE (Default: TRUE).}
}
\description{
Display only a single taxon from a phyloseq object
}
\examples{
plotSingleTaxon(physeq, "ASV0001", "group", colorpal = myPalette21)
plotSingleTaxon(physeq, "Pseudomonas", "group", taxlevel = "Genus", colorpal = myPalette21)
plotSingleTaxon(physeq, "Firmicutes", "group", "timepoint", taxlevel = "Phylum", colorpal = myPalette21)
}
