% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRNA_copy_number_correction.R
\name{rRNA_copy_number_correction}
\alias{rRNA_copy_number_correction}
\title{Correct OTU abundances according to rRNA copy numbers}
\usage{
rRNA_copy_number_correction(physeq, dbpath, use_median = F)
}
\arguments{
\item{physeq}{A phyloseq object containg at least abundance (otu_table) and taxonomy (tax_table) data.}

\item{dbpath}{Path to the rrndb file.}

\item{median}{if true, the median copy number is used on the genus level; if FALSE, mean is used for all tax levels (default)}
}
\value{
phyloseq object with copy-number corrected OTU counts
}
\description{
THis script uses information from the rrnDB (https://rrndb.umms.med.umich.edu/).
Each taxonomic rank present in the input data is screened for taxa present in the data base.
If the name is found, the mean or median (depending on parameters) is used to normalize the read counts for this taxon.
last changed: A. Dötsch, 21.12.2021
}
\examples{
rrnDB_path <- "/home/bioinf/scripts/rrnDB-5.7_pantaxa_stats_RDP_Silva.tsv"
physeq_norm <- rRNA_copy_number_correction(physeq, dbpath = rrnDB_path)

}
