% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxLevelDetermination.R
\name{taxLevelDetermination}
\alias{taxLevelDetermination}
\title{Calculate level of determination of for each taxonomic level}
\usage{
taxLevelDetermination(taxtab)
}
\arguments{
\item{taxtab}{Taxonomy table - either a data.frame or phyloseq object created with tax_table()}
}
\value{
tld
}
\description{
Counts the number of NA assigned to sequence variants at each level and devides by the total number of ASVs.
}
