% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackReadsDada.R
\name{trackReadsDada}
\alias{trackReadsDada}
\title{Displays amount of reads at different steps during dada2 processing}
\usage{
trackReadsDada(
  filtered,
  denoisedF,
  denoisedR,
  merged,
  nochim,
  relative = T,
  graphicOutput = T
)
}
\arguments{
\item{filtered}{Output table generate by filterAndTrim()}

\item{denoisedF}{output of dada() after denoising forward reads}

\item{denoisedR}{output of dada() after denoising reverse reads}

\item{merged}{output of mergePairs() after merging denoised forward and reverse reads}

\item{nochim}{output of removeBimeraDenovo() after removing chimeras}

\item{relative}{display results relative to initial number of reads (default: T)}
}
\value{
track
}
\description{
Inputs are data objects at various steps in the dada2 workflow and this script summarizes the amounts of reads that remain at each step.
Results are displayed as boxplots.
}
